Split out identifier-related code
